+++ paddle_billing/Entities/Notifications/__init__.py
@@ -1,3 +1,4 @@
+from paddle_billing.Entities.Notifications.NotificationEvent import NotificationEvent
 from paddle_billing.Entities.Notifications.NotificationOrigin import NotificationOrigin
 from paddle_billing.Entities.Notifications.NotificationPayout import NotificationPayout
 from paddle_billing.Entities.Notifications.NotificationPayoutStatus import NotificationPayoutStatus