+++ paddle_billing/Notifications/Verifier.py
@@ -1,10 +1,9 @@
-import requests
-
 from os import getenv
 from time import time
 
 from paddle_billing.Logger import get_logger
 
+from paddle_billing.Notifications.Requests.Request import Request
 from paddle_billing.Notifications.Secret import Secret
 from paddle_billing.Notifications.PaddleSignature import PaddleSignature
 
@@ -20,7 +19,7 @@ class Verifier:
     def maximum_variance(self) -> int:
         return self.__maximum_variance
 
-    def verify(self, request: requests, secrets: list[Secret] | Secret, verify_time_drift=True):
+    def verify(self, request: Request, secrets: list[Secret] | Secret, verify_time_drift=True):
         """
         :param request:           The request object to verify
         :param secrets:           One or more Secrets to use for verifying the request