+++ tests/Utils/__init__.py
@@ -1,1 +1,1 @@
-from .FakeResponse import FakeResponse
+from .FakeRequest import FakeRequest